amd-xgbe: add support for sfp BASE-BX10
amd-xgbe is missing support for SFP Modules with tranceiver type,
Ethernet: BASE-BX10, or other sfp modules with fibre.
With this patch it will recognize BX10 and sfp modules based on
wavelength if BX10 is not defined.

Autoneg is enabled by default because some ISP requires this to
receive an IP with DHCP.

Tested on FS, Ubiquiti and Pro Optix sfp modules, BiDi 1310/1550 SC/LC.
Note: Added back min/max for sfp_speed or it would not be able
to detect the correct speed.

Signed-off-by: Kristian Skramstad <kristian+github@83.no>